+++ enterprise/cmd/frontend/internal/licensing/enforcement.go
@@ -39,7 +39,7 @@ func NewPreCreateUserHook(s UsersStore) func(context.Context) error {
 			} else {
 				message := "Unable to create user account: "
 				if info == nil {
-					message = fmt.Sprintf("a Sourcegraph subscription is required to exceed %d users. A site admin must purchase a subscription at https://sourcegraph.com/user/subscriptions/new. Enter the license key in the Sourcegraph management console.", NoLicenseMaximumAllowedUserCount)
+					message = fmt.Sprintf("a Sourcegraph subscription is required to exceed %d users. A site admin must purchase a subscription at https://sourcegraph.com/user/settings/subscriptions/new. Enter the license key in the Sourcegraph management console.", NoLicenseMaximumAllowedUserCount)
 				} else {
 					message += "the Sourcegraph subscription's maximum user count has been reached. A site admin must upgrade the Sourcegraph subscription to allow for more users. Enter the license key in the Sourcegraph management console (https://docs.sourcegraph.com/admin/management_console)."
 				}