+++ web/src/platform/context.ts
@@ -1,4 +1,4 @@
-import { concat, ReplaySubject } from 'rxjs'
+import { concat, Observable, ReplaySubject } from 'rxjs'
 import { map, publishReplay, refCount } from 'rxjs/operators'
 import { createExtensionHost } from '../../../shared/src/api/extension/worker'
 import { gql } from '../../../shared/src/graphql/graphql'
@@ -6,11 +6,11 @@ import * as GQL from '../../../shared/src/graphql/schema'
 import { PlatformContext } from '../../../shared/src/platform/context'
 import { mutateSettings, updateSettings } from '../../../shared/src/settings/edit'
 import { gqlToCascade } from '../../../shared/src/settings/settings'
+import { createAggregateError } from '../../../shared/src/util/errors'
 import { LocalStorageSubject } from '../../../shared/src/util/LocalStorageSubject'
 import { toPrettyBlobURL } from '../../../shared/src/util/url'
-import { requestGraphQL } from '../backend/graphql'
+import { queryGraphQL, requestGraphQL } from '../backend/graphql'
 import { Tooltip } from '../components/tooltip/Tooltip'
-import { fetchViewerSettings } from '../user/settings/backend'
 
 /**
  * Creates the {@link PlatformContext} for the web app.
@@ -68,4 +68,58 @@ export function createPlatformContext(): PlatformContext {
         sideloadedExtensionURL: new LocalStorageSubject<string | null>('sideloadedExtensionURL', null),
     }
     return context
+}
+
+const settingsCascadeFragment = gql`
+    fragment SettingsCascadeFields on SettingsCascade {
+        subjects {
+            __typename
+            ... on Org {
+                id
+                name
+                displayName
+            }
+            ... on User {
+                id
+                username
+                displayName
+            }
+            ... on Site {
+                id
+                siteID
+            }
+            latestSettings {
+                id
+                contents
+            }
+            settingsURL
+            viewerCanAdminister
+        }
+        final
+    }
+`
+
+/**
+ * Fetches the viewer's settings from the server. Callers should use settingsRefreshes#next instead of calling
+ * this function, to ensure that the result is propagated consistently throughout the app instead of only being
+ * returned to the caller.
+ *
+ * @return Observable that emits the settings
+ */
+function fetchViewerSettings(): Observable<GQL.ISettingsCascade> {
+    return queryGraphQL(gql`
+        query ViewerSettings {
+            viewerSettings {
+                ...SettingsCascadeFields
+            }
+        }
+        ${settingsCascadeFragment}
+    `).pipe(
+        map(({ data, errors }) => {
+            if (!data || !data.viewerSettings) {
+                throw createAggregateError(errors)
+            }
+            return data.viewerSettings
+        })
+    )
 }