+++ web/src/user/settings/emails/AddUserEmailForm.tsx
@@ -2,12 +2,12 @@ import { upperFirst } from 'lodash'
 import * as React from 'react'
 import { merge, Observable, of, Subject, Subscription } from 'rxjs'
 import { catchError, map, switchMap, tap } from 'rxjs/operators'
-import { gql } from '../../../../shared/src/graphql/graphql'
-import * as GQL from '../../../../shared/src/graphql/schema'
-import { createAggregateError, ErrorLike } from '../../../../shared/src/util/errors'
-import { mutateGraphQL } from '../../backend/graphql'
-import { Form } from '../../components/Form'
-import { eventLogger } from '../../tracking/eventLogger'
+import { gql } from '../../../../../shared/src/graphql/graphql'
+import * as GQL from '../../../../../shared/src/graphql/schema'
+import { createAggregateError, ErrorLike } from '../../../../../shared/src/util/errors'
+import { mutateGraphQL } from '../../../backend/graphql'
+import { Form } from '../../../components/Form'
+import { eventLogger } from '../../../tracking/eventLogger'
 
 interface Props {
     /** The GraphQL ID of the user with whom the new emails are associated. */